requirements.txt
gradio>=4.0
python-dotenv>=1.0.0
# OpenAI Agents SDK (already in your project)
openai-agents
markdown>=3.0
